+++ internal/clioptions/clioptions.go
@@ -149,7 +149,7 @@ func (o *CLIOptions) AddJWTServiceAccountFlags(flags *pflag.FlagSet) {
 }
 
 func (o *CLIOptions) AddEditServiceAccountFlags(flags *pflag.FlagSet) {
-	flags.StringVarP(&o.IAMRole, "role", "r", "", "the new company role for the user")
+	flags.StringVarP(&o.IAMRole, "role", "r", "", "the new company role for the service account")
 	flags.StringVarP(&o.ServiceAccountID, "service-account-id", "", "", "the service account id to edit")
 }
 
@@ -170,6 +170,11 @@ func (o *CLIOptions) CreateNewGroupFlags(flags *pflag.FlagSet) {
 func (o *CLIOptions) AddMemberToGroupFlags(flags *pflag.FlagSet) {
 	flags.StringSliceVarP(&o.UserEmails, "user-email", "", []string{}, "the list of user email to add to the group")
 	flags.StringVarP(&o.GroupID, "group-id", "", "", "the group id where to add the users")
+}
+
+func (o *CLIOptions) AddEditGroupFlags(flags *pflag.FlagSet) {
+	flags.StringVarP(&o.IAMRole, "role", "r", "", "the new company role for the group")
+	flags.StringVarP(&o.GroupID, "group-id", "", "", "the group id to edit")
 }
 
 func (o *CLIOptions) AddMarketplaceApplyFlags(cmd *cobra.Command) {