+++ internal/cmd/company/iam/serviceaccount/basic/basic.go
@@ -45,7 +45,7 @@ service account is created on the company.`,
 			cobra.CheckErr(err)
 			client, err := client.APIClientForConfig(restConfig)
 			cobra.CheckErr(err)
-			credentials, err := createBasicServiceAccount(cmd.Context(), client, serviceAccountName, restConfig.CompanyID, resources.ServiceAccountRole(options.IAMRole))
+			credentials, err := createBasicServiceAccount(cmd.Context(), client, serviceAccountName, restConfig.CompanyID, resources.IAMRole(options.IAMRole))
 			if err != nil {
 				return err
 			}
@@ -59,16 +59,7 @@ service account is created on the company.`,
 
 	// add cmd flags
 	options.AddServiceAccountFlags(cmd.Flags())
-	err := cmd.RegisterFlagCompletionFunc("role", func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		return []string{
-			resources.ServiceAccountRoleGuest.String(),
-			resources.ServiceAccountRoleReporter.String(),
-			resources.ServiceAccountRoleDeveloper.String(),
-			resources.ServiceAccountRoleMaintainer.String(),
-			resources.ServiceAccountRoleProjectAdmin.String(),
-			resources.ServiceAccountRoleCompanyOwner.String(),
-		}, cobra.ShellCompDirectiveDefault
-	})
+	err := cmd.RegisterFlagCompletionFunc("role", resources.IAMRoleCompletion)
 
 	if err != nil {
 		// we panic here because if we reach here, something nasty is happening in flag autocomplete registration
@@ -78,8 +69,8 @@ service account is created on the company.`,
 	return cmd
 }
 
-func createBasicServiceAccount(ctx context.Context, client *client.APIClient, name, companyID string, role resources.ServiceAccountRole) ([]string, error) {
-	if !resources.IsValidServiceAccountRole(role) {
+func createBasicServiceAccount(ctx context.Context, client *client.APIClient, name, companyID string, role resources.IAMRole) ([]string, error) {
+	if !resources.IsValidIAMRole(role) {
 		return nil, fmt.Errorf("invalid service account role %s", role)
 	}